5. 유전 시스템, 6. 종족유전 알림 구현 (버그 미테스트)

+++ Project-FW/GameScene.cpp
@@ -35,7 +35,7 @@ Scene* GameScene::scene()
 
 void GameScene::Init()
 {
-	float Height=g_D3dDevice->GetWinHeight() ;
+	float Height = (float)g_D3dDevice->GetWinHeight() ;
 
 	g_CameraManager->AddCamera(new CCamera(), 0) ;
